refactored bot for more modular code
