+++ server.js
@@ -37,7 +37,7 @@ instrument(io, {
 
 // Room_services
 const { createRoom } = require('./services/room_services/createRoom.service.js');
-const { slackApp } = require('./bot.js');
+const { slackApp } = require('./bot/bot.js');
 const { auto_refresh } = require('./services/auth_services/refresh_token.service.js');
 require('dotenv').config;
